退款类型转换

+++ xiaoerke-order/xiaoerke-order-service/src/main/java/com/cxqm/xiaoerke/modules/order/service/impl/ConsultPhonePatientServiceImpl.java
@@ -162,7 +162,7 @@ public class ConsultPhonePatientServiceImpl implements ConsultPhonePatientServic
 
     @Override
     @Transactional(rollbackFor=Exception.class)
-    public BigDecimal cancelOrder(Integer phoneConsultaServiceId,String cancelReason,String cancelState) throws CancelOrderException {
+    public Float cancelOrder(Integer phoneConsultaServiceId,String cancelReason,String cancelState) throws CancelOrderException {
         int sysOrderState = 0;
 //        Float price = 0f;
         //取消订单
@@ -210,7 +210,7 @@ public class ConsultPhonePatientServiceImpl implements ConsultPhonePatientServic
                 throw  new CancelOrderException();
         }
 
-        return (BigDecimal)map.get("price");
+        return (Float)map.get("price");
     }
 
     /**
